// client/src/app/api/openapi/backend/model/ort.ts
/**
 * Messe Server API
 * API für die Messe-Anwendung
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Ort = 'Stand' | 'Lager';

export const Ort = {
    Stand: 'Stand' as Ort,
    Lager: 'Lager' as Ort
};